requirements.txt
streamlit
streamlit-js-eval
pillow
altair
pandas